More Monero integrated address fixes

+++ src/MiningCore/Blockchain/Monero/MoneroConstants.cs
@@ -18,6 +18,7 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
+using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Text.RegularExpressions;
@@ -38,35 +39,35 @@ namespace MiningCore.Blockchain.Monero
     {
         public const string WalletDaemonCategory = "wallet";
 
-        public static readonly Dictionary<CoinType, int> AddressLength = new Dictionary<CoinType, int>
+        public static readonly Dictionary<CoinType, UInt64> AddressLength = new Dictionary<CoinType, UInt64>
         {
             { CoinType.XMR, 95 },
             { CoinType.ETN, 98 },
             { CoinType.AEON, 97 },
         };
 
-        public static readonly Dictionary<CoinType, int> AddressPrefix = new Dictionary<CoinType, int>
+        public static readonly Dictionary<CoinType, UInt64> AddressPrefix = new Dictionary<CoinType, UInt64>
         {
             { CoinType.XMR, 18 },
             { CoinType.ETN, 18018 },
             { CoinType.AEON, 178 },
         };
 
-        public static readonly Dictionary<CoinType, int> AddressPrefixTestnet = new Dictionary<CoinType, int>
+        public static readonly Dictionary<CoinType, UInt64> AddressPrefixTestnet = new Dictionary<CoinType, UInt64>
         {
             { CoinType.XMR, 53 },
             { CoinType.ETN, 53 },
             { CoinType.AEON, 178 },
         };
 
-        public static readonly Dictionary<CoinType, int> AddressPrefixIntegrated = new Dictionary<CoinType, int>
+        public static readonly Dictionary<CoinType, UInt64> AddressPrefixIntegrated = new Dictionary<CoinType, UInt64>
         {
             { CoinType.XMR, 19 },
             { CoinType.ETN, 18019 },
             { CoinType.AEON, 178 },
         };
 
-        public static readonly Dictionary<CoinType, int> AddressPrefixIntegratedTestnet = new Dictionary<CoinType, int>
+        public static readonly Dictionary<CoinType, UInt64> AddressPrefixIntegratedTestnet = new Dictionary<CoinType, UInt64>
         {
             { CoinType.XMR, 54 },
             { CoinType.ETN, 54 },
